PlayerHealth
Handles respawns, health pickups, healing and health UI indicator.
